Add ability to set image based on rules; update tests, example
The rules array isn't well defined yet -- possibly it needs
to be made up of some new ImageRules class, but for the
moment this seemed to be the simplest.

Also the agent rendering will need to be updated to
deal with changing properties -- at the moment the
image that is used when an agent is first rendered is the one
it continues to use.