Compile .coffee in interactives/ **requires npm install**
This is so that we can author individual models in CS in the interactives folder and compile it to individual JS files in 
public/.

There doesn't seem to be a native brunch way to do this, so 
this uses the (kinda hacky) after-brunch plugin to execute 
arbitrary commands after each compile.

Since the .coffee files were already being brought over from
interactives/ intact (since interactives/ is now the equivalent
of assets/) it would have been nice to simply compile them
in-place. However, the coffee compile doesn't seem to be able
to do this. Further, when I tried to compile and then delete

    coffee --compile public
    find public/ -type f -name "*.coffee" -delete

the delete command would happen faster than the compile command. 
The only solution seems to be to remove the existing .coffee 
files in public/ and compile the ones in app/.