Add first two models to their own files
This is slightly strawman so that we can continue to think about
how we want to refactor the models. The models -one-plant and
-three-plants are 99% identical.

One possibility is to load in the environment+rules from another
file. Is this too many separate files for authors?

Also a question is whether there is any good way to deal with the
html files -- currently they are identical except for the model
script being loaded. Should we instead of a single model.html file,
and point to a script with a hash in the url?