Make more properties species-level props
Agent properties should just be for props that change within a species,
props is "is annual" should be at the species level.

This will improve efficiency because these props won't need to be
copied to offspring, and lets us set model-level differences (like
changing a plant to not be able to seed) just once, instead of
in each addOrganism button as we were doing.