Keep agent.speed constant, pass desired speed to move()
Previous version was modifying speed every time wander() or other
methods were being called. The result of this was that if we called
wander() too many times in a row, the agent would slow down to a crawl.

Here we keep @speed as a property, but just move a desired amount by
passing a speed to move.