Add ability to define animated sprites for agent images
Agents can now define an animated sprite with one or more
named animations in place of a regular layer image.

See next commit for example of this.

Note that the animated-agent's movements is just an initial pattern
for how we can call the individual animations. It may change in the
future.