Make BasicAnimals change direction in smaller increments
Now that we can visualize the changing direction of an agent (see
hawks and rabbits), the sudden direction changes look ugly.

This makes it so that a random wander will change direction by a smaller
amount each time, and a directed movement (movement towards a target)
will change direction by 1/20th of the angle each time.

Note: We should make the random wander nicer than a random walk, which
doesn't look very natural (e.g. pick a spot and move to it.)