Can now set base environment width.
We can now independently set the desired "base" environment width, which
is the width we would see if we didn't then apply any scaling, which is
what `setEnvironmentDisplayWidth` does.

This is useful because setting the display width via scaling applies a
css transform, which can reduce the resolution of images, even if we are
starting with high-res images.

If `viewWidth / viewHeight` are not set, the view will default to the
model width and height, as usual.